# Optimize all transactions query