- Don't use a join for multisig hashes and module hashes. Check is faster when using the actual tables with their indexes
- Use `UNION ALL` instead of `UNION`. There are already checks for duplicated elements
- Force a pre-sort of tables by timestamp DESC. That way PostgreSQL optimizer will use a heapsort instead of quicksort

Before:
```
Disk: 643080kB (for ERC20)
Planning Time: 1.834 ms
Execution Time: 20388.061 ms
```
After:
```
Disk usage: 154936kB (for ERC20) 
Planning Time: 1.904 ms
Execution Time: 6629.214 ms
```